backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
numpy<2
pandas==2.2.3
opencv-python==4.10.0.84
pydantic==2.11.4
pydantic_core==2.33.2
scikit-learn==1.5.2
scipy==1.14.1
python-dotenv==1.1.0
PyYAML==6.0.3
starlette==0.46.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.11.8
uvicorn==0.34.2
watchfiles==1.1.1
websockets==16.0
earthengine-api
google-auth
